Fix: Integer conversion warning, formatting

+++ test/src/test_random_annex_xyz.cpp
@@ -45,7 +45,8 @@ inline void invoke_max_annex_size(uint32_t max_symbols,
 {
     Partitioning scheme(max_symbols, max_symbol_size, object_size);
 
-    uint32_t max_annex = kodo::max_annex_size(max_symbols, max_symbol_size, object_size);
+    uint32_t max_annex = kodo::max_annex_size(
+        max_symbols, max_symbol_size, object_size);
 
     for(uint32_t i = 0; i < scheme.blocks(); ++i)
     {
@@ -69,7 +70,7 @@ TEST(TestRandomAnnexBase, max_annex)
 //
 // Tests encoding and decoding with a random annex wrapped
 // encoder / decoder
-// 
+//
 template
 <
     class Encoder,
@@ -84,34 +85,40 @@ void invoke_random_annex_partial(uint32_t max_symbols,
     uint32_t object_size = max_symbols * max_symbol_size * multiplier;
     object_size -= (rand() % object_size);
 
-    uint32_t annex_size = kodo::max_annex_size(max_symbols, max_symbol_size, object_size);
+    uint32_t annex_size = kodo::max_annex_size(max_symbols,
+                                               max_symbol_size, object_size);
 
     if(annex_size > 0)
     {
         // Randomize the actual annex size
         annex_size -= (rand() % annex_size);
     }
-    
-    typedef kodo::random_annex_encoder<Encoder, Partitioning> random_annex_encoder;
-    typedef kodo::random_annex_decoder<Decoder, Partitioning> random_annex_decoder;
-    
+
+    typedef kodo::random_annex_encoder<Encoder, Partitioning>
+        random_annex_encoder;
+
+    typedef kodo::random_annex_decoder<Decoder, Partitioning>
+        random_annex_decoder;
+
     std::vector<char> data_in(object_size);
     std::vector<char> data_out(object_size, '\0');
 
     kodo::random_uniform<char> fill_data;
     fill_data.generate(&data_in[0], data_in.size());
-    
+
     typename Encoder::factory encoder_factory(max_symbols, max_symbol_size);
     typename Decoder::factory decoder_factory(max_symbols, max_symbol_size);
 
     //std::cout << "object size " << data_in.size() << std::endl;
     //std::cout << "annex size " << annex_size << std::endl;
-    
-    random_annex_encoder obj_encoder(annex_size, encoder_factory, kodo::storage(data_in));
+
+    random_annex_encoder obj_encoder(
+        annex_size, encoder_factory, kodo::storage(data_in));
 
     //std::cout << "encoders " << obj_encoder.encoders() << std::endl;
-    
-    random_annex_decoder obj_decoder(annex_size, decoder_factory, obj_encoder.object_size());
+
+    random_annex_decoder obj_decoder(
+        annex_size, decoder_factory, obj_encoder.object_size());
 
     EXPECT_TRUE(obj_encoder.encoders() >= 1);
     EXPECT_TRUE(obj_decoder.decoders() >= 1);
@@ -119,12 +126,13 @@ void invoke_random_annex_partial(uint32_t max_symbols,
 
     uint32_t bytes_used = 0;
 
-    
-    
+
+
     for(uint32_t i = 0; i < obj_encoder.encoders(); ++i)
     {
         typename Encoder::pointer encoder = obj_encoder.build(i);
-        typename random_annex_decoder::pointer_type decoder = obj_decoder.build(i);
+        typename random_annex_decoder::pointer_type decoder =
+            obj_decoder.build(i);
 
         encoder->systematic_off();
 
@@ -134,31 +142,32 @@ void invoke_random_annex_partial(uint32_t max_symbols,
         EXPECT_TRUE(encoder->block_size() == decoder->block_size());
         EXPECT_TRUE(encoder->bytes_used() == decoder->bytes_used());
         EXPECT_TRUE(encoder->payload_size() == decoder->payload_size());
-        
+
         std::vector<uint8_t> payload(encoder->payload_size());
-        
+
         while(!decoder->is_complete())
         {
             encoder->encode( &payload[0] );
             decoder->decode( &payload[0] );
         }
-        
+
         kodo::mutable_storage storage = kodo::storage(
             &data_out[0] + bytes_used, decoder->bytes_used());
-        
+
         kodo::copy_symbols(storage, decoder.unwrap());
-        
+
         bytes_used += decoder->bytes_used();
     }
-    
+
     EXPECT_EQ(bytes_used, object_size);
     EXPECT_TRUE(std::equal(data_in.begin(), data_in.end(), data_out.begin()));
-    
+
 }
 
 
 
-void test_random_annex_coders(uint32_t symbols, uint32_t symbol_size, uint32_t multiplier)
+void test_random_annex_coders(uint32_t symbols, uint32_t symbol_size,
+                              uint32_t multiplier)
 {
 
     invoke_random_annex_partial<
@@ -170,7 +179,7 @@ void test_random_annex_coders(uint32_t symbols, uint32_t symbol_size, uint32_t m
         kodo::full_rlnc8_encoder,
         kodo::full_rlnc8_decoder,
             kodo::rfc5052_partitioning_scheme>(symbols, symbol_size, multiplier);
-    
+
     invoke_random_annex_partial<
         kodo::full_rlnc16_encoder,
         kodo::full_rlnc16_decoder,
@@ -182,17 +191,17 @@ void test_random_annex_coders(uint32_t symbols, uint32_t symbol_size, uint32_t m
 TEST(TestRandomAnnexCoder, construct_and_invoke_the_basic_api)
 {
     test_random_annex_coders(32, 1600, 2);
-    
-    srand(time(0));
-    
+
+    srand(static_cast<uint32_t>(time(0)));
+
     uint32_t symbols = (rand() % 256) + 1;
     uint32_t symbol_size = ((rand() % 2000) + 1) * 2;
-    
+
     // Multiplies the data to be encoded so that the object encoder
     // is expected to contain multiplier encoders.
     uint32_t multiplier = (rand() % 10) + 1;
-    
-    test_random_annex_coders(symbols, symbol_size, multiplier);  
+
+    test_random_annex_coders(symbols, symbol_size, multiplier);
 }
